Added history charts to zone overview tiles
