+++ libs/charts/src/lib/components/chart-container/chart-container.component.html
@@ -1,2 +1,1 @@
 <canvas #chartContainer width="100%" height="100%"></canvas>
-
