+++ libs/ues/effluent/src/lib/modules/core/components/campus-overview-list/campus-overview-list.component.html
@@ -1,3 +1,3 @@
 <ng-container *ngFor="let z of zones | async">
-  <tamu-gisc-zone-overview-tile [zone]="z" (click)="highlightZone(z.attributes.SampleNumber)"></tamu-gisc-zone-overview-tile>
+  <tamu-gisc-zone-overview-tile [zone]="z"></tamu-gisc-zone-overview-tile>
 </ng-container>
