+++ libs/ues/effluent/src/lib/modules/core/components/site-history-chart/site-history-chart.component.scss
@@ -1,0 +1,3 @@
+:host {
+  width: 100%;
+}
