+++ libs/ues/effluent/src/lib/modules/core/components/site-history-chart/site-history-chart.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, OnInit, Input, HostBinding } from '@angular/core';
 import { Observable } from 'rxjs';
 
 import { IChartConfiguration } from '@tamu-gisc/charts';
