+++ libs/ues/effluent/src/lib/modules/core/components/zone-overview-tile/zone-overview-tile.component.html
@@ -1,16 +1,22 @@
 <div class="zone-overview-tile">
-  <div class="zone-metric">
-    <div class="zone-metric-value">{{zone.attributes.SampleNumber}}</div>
-    <div class="zone-metric-name">Sample</div>
+  <div class="zone-overview-row" (click)="handleToggle()">
+    <div class="zone-metric">
+      <div class="zone-metric-value">{{zone.attributes.SampleNumber}}</div>
+      <div class="zone-metric-name">Sample</div>
+    </div>
+
+    <div class="zone-metric">
+      <div class="zone-metric-value">{{zone.attributes.Focus | number}}</div>
+      <div class="zone-metric-name">Focus Area (Sq. Ft.)</div>
+    </div>
+
+    <div class="zone-metric">
+      <div class="zone-metric-value" [ngClass]="{'metric-safe': zone?.attributes?.Signal?.result <= .45, 'metric-warning': zone?.attributes?.Signal?.result > .45, 'metric-alert' : zone?.attributes?.Signal?.result > .6}">{{zone?.attributes?.Signal?.result ? zone.attributes.Signal.result : 'N/A'}}</div>
+      <div class="zone-metric-name">Signal</div>
+    </div>
   </div>
 
-  <div class="zone-metric">
-    <div class="zone-metric-value">{{zone.attributes.Focus | number}}</div>
-    <div class="zone-metric-name">Focus Area (Sq. Ft.)</div>
-  </div>
-
-  <div class="zone-metric">
-    <div class="zone-metric-value" [ngClass]="{'metric-safe': zone?.attributes?.Signal?.result <= .45, 'metric-warning': zone?.attributes?.Signal?.result > .45, 'metric-alert' : zone?.attributes?.Signal?.result > .6}">{{zone?.attributes?.Signal?.result ? zone.attributes.Signal.result : 'N/A'}}</div>
-    <div class="zone-metric-name">Signal</div>
+  <div class="zone-overview-row" *ngIf="toggled | async">
+    <tamu-gisc-site-history-chart [tier]="tier" [sample]="sample"></tamu-gisc-site-history-chart>
   </div>
 </div>
